Consolidated Aura Generation, resprited Zephyr stuff

+++ Items/Accessories/MicitBangle.cs
@@ -24,7 +24,7 @@ namespace ProvidenceMod.Items.Accessories
     {
       player.Providence().micitBangle = true;
       player.Providence().parryCapable = true;
-      player.Providence().parryType = ParryTypeID.Support;
+      player.Providence().parryType = (int)ParryTypeID.Support;
       player.Providence().hemomancy = true;
     }
     public override void AddRecipes()
